Swig: Move repository implementation from rpm to new repo module
