+++ bindings/CMakeLists.txt
@@ -43,7 +43,7 @@ endif()
 
 
 # list of all modules that will be included in libdnf bindings
-list(APPEND SWIG_LIBDNF_MODULES base common comps conf logger rpm transaction plugin)
+list(APPEND SWIG_LIBDNF_MODULES base common comps conf logger repo rpm transaction plugin)
 
 
 # list of all modules that will be included in libdnf-cli bindings
