+++ test/python3/libdnf/repo/test_repo.py
@@ -15,43 +15,45 @@
 # You should have received a copy of the GNU Lesser General Public License
 # along with libdnf.  If not, see <https://www.gnu.org/licenses/>.
 
-require 'test/unit'
-require 'tmpdir'
-include Test::Unit::Assertions
+import unittest
+import tempfile
+import shutil
+import os
 
-require 'libdnf/base'
+import libdnf.base
+import libdnf.logger
 
-class TestRepo < Test::Unit::TestCase
-    def test_repo()
-        base = Base::Base.new()
+
+class TestRepo(unittest.TestCase):
+    def test_repo(self):
+        base = libdnf.base.Base()
 
         # Sets path to cache directory.
-        tmpdir = Dir.mktmpdir("libdnf-ruby-")
-        base.get_config().cachedir().set(Conf::Option::Priority_RUNTIME, tmpdir)
+        tmpdir = tempfile.mkdtemp(prefix="libdnf-python3-")
+        base.get_config().cachedir().set(libdnf.conf.Option.Priority_RUNTIME, tmpdir)
 
-        repo_sack = Rpm::RepoSack.new(base)
-        package_sack = Rpm::PackageSack.new(base)
+        repo_sack = libdnf.repo.RepoSack(base)
+        package_sack = libdnf.rpm.PackageSack(base)
 
         # Creates system repository and loads it into rpm::PackageSack.
-        package_sack.create_system_repo(false)
+        package_sack.create_system_repo(False)
 
         # Creates new repositories in the repo_sack
         repo = repo_sack.new_repo("repomd-repo1")
 
         # Tunes repository configuration (baseurl is mandatory)
-        repo_path = File.join(Dir.getwd(), '../../../test/data/repos-repomd/repomd-repo1/')
-        baseurl = 'file://' + repo_path
+        repo_path = os.path.join(os.getcwd(), "../../../test/data/repos-repomd/repomd-repo1/")
+        baseurl = "file://" + repo_path
         repo_cfg = repo.get_config()
-        repo_cfg.baseurl().set(Conf::Option::Priority_RUNTIME, baseurl)
+        repo_cfg.baseurl().set(libdnf.conf.Option.Priority_RUNTIME, baseurl)
 
         # Loads repository into rpm::Repo.
         repo.load()
 
         # Loads rpm::Repo into rpm::PackageSack
-        package_sack.load_repo(repo.get(), Rpm::PackageSack::LoadRepoFlags_USE_PRESTO |
-                            Rpm::PackageSack::LoadRepoFlags_USE_UPDATEINFO | Rpm::PackageSack::LoadRepoFlags_USE_OTHER)
+        PackageSack = libdnf.rpm.PackageSack
+        package_sack.load_repo(repo.get(), PackageSack.LoadRepoFlags_USE_PRESTO | PackageSack.LoadRepoFlags_USE_UPDATEINFO |
+                            PackageSack.LoadRepoFlags_USE_OTHER)
 
         # Remove the cache directory.
-        FileUtils.remove_entry(tmpdir)
-    end
-end
+        shutil.rmtree(tmpdir)
